added Midi + Midi-clock sync removed Random Mode
Midi-clock was implemented using a modified version of beatclock at the time. Since then, there has been an update for norns that added more global clocking features. Reimplementing this is therefore likely in the future.

Random Mode was removed because the change to parameters was not instantaneous, so samples would "slide" between different values. Additionally, sometimes the parameters would line up in such a way that the played sample was incredibly loud and harsh.